wbtest/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
et-xmlfile==1.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
xlrd==2.0.1
yarl==1.7.2
